--- a/src/hnsw.rs
+++ b/src/hnsw.rs
@@ -28,10 +28,10 @@ use log::trace;
 pub use crate::filter::FilterT;
 use anndists::dist::distances::Distance;
 
-#[cfg(feature = "mimalloc")]
+#[cfg(feature = "use_mimalloc")]
 use mimalloc::MiMalloc;
 
-#[cfg(feature = "mimalloc")]
+#[cfg(feature = "use_mimalloc")]
 #[global_allocator]
 static GLOBAL: MiMalloc = MiMalloc;