@mexus 2024-08-21
As per https://doc.rust-lang.org/stable/std/alloc/index.html#the-global_allocator-attribute:

> The #[global_allocator] can only be used once in a crate or its recursive dependencies.

I would not recommend setting the global allocator here.

Normally the global allocator is selected by the end-user, i.e. in executables. Changing the global allocator in a library is an anti-pattern though, since it implicitly leak into the whole application through any levels of transitive dependencies. For example, if I create a library crate that uses the `annembed` crate, which depends on this (`hnsw_rs`) crate, enabling non-standard allocator here will implicitly enable it for all the users of "my" library! Enabling it through a feature helps a bit, but doesn't make a lot of sense, since the end user can simply select the preferred allocator anyhow.

In library, it might make sense to enable some non-standard global allocator specifically in benchmarks though. This way dependents of this library won't get their allocator suddenly changed :)